Improve MEX writer
Add support for auto flushing when writing MEX files. This is only used
when writing to a directory. Also, only flush when used from the CLI.

Compress entries in a MEX stream to minimize memory usage.
Unfortunately, we need to know the size of an entry before writing it,
so we might as well compress it to reduce memory usage.