Add a fast method for mapping FVs to BMs
Use this for building the design matrix for DEA and printing the
experimental design and cell-level metadata.

Inline getOrderedFactors() in favour of simply using the
ExperimentalFactor.COMPARATOR when sorting factors is desired.

Relocate constants relevant to generating design matrix in
DiffExAnalyzerUtils.

Relocate batch factor-related utilities in ExperimentalFactorUtils.

Add tests for unassigned samples and samples with more than one FV
assigned.