Improve warnings for unrecognized nodes in PubMedXMLParser
Reorder methods to be more readable. Always put the Node or NodeList
being processed as the first parameter.

Don't check for an ERROR tag, it's already done in
EntrezXmlUtils.parse().

Reuse parsing of date when populating the publication date.

Always use warnings and indicate the name of the parent node.