Detect the best font to use for rendering diagnostic plots
It is practically impossible to perform this from the browser with
JavaScript, so instead rely on the User-Agent string to detect major OS.
The font is detected from the HTML page and then passed on to the links
that are generated. Arial is used as a default (with a fallback to
Liberation Sans).

Pass the font to the frontend via the detailed EE VO and JSP tag
attributes.

Eliminate DPI in the PNG metadata, it is meaningless.

Fill the creation date metadata.