Remove unapplicable and misleading @Secured annotation on controllers
The @Secured (and also @Transactional) annotations are not applied to
controllers because those reside in the a servlet-based Spring context
that is not subject to post-processing declared in the core context.

Add a @ContextHierarchy to the BaseWebIntegrationTest to reproduce this
behavior by separating the core context from the servlet context.