Few improvements for DEA CLI
Allow interactions to be specified by the -factors argument using the
'a:b' syntax.

Do not include continuous factors in interactions.

Allow automatic selection of factors when a subset factor is passed.

Replace redundant getAnalyes() with findByExperiment().

Add a --complete-factors option to provide completions for factors and
interactions. This is not fully integrated because we do not presently
have a mechanism to recover the EE identifier, but it can be used to
discover factors.