Prototype for filtering 10x MEX data
Add detection for unfiltered 10x data

The MTX file contains JSON metadata that can be used to detect if Cell
Ranger was used to generate the file.

In addition, we can safely assume that the presence of unused barcodes
is an indicator of an unfiltered dataset. If that does not apply in
real world, we could also use a threshold on the unused fraction.