Allow for parallel filtering of 10x MEX data
Add a -transformThreads/--transform-threads option to
downloadSingleCellData and loadSingleCellData to control the number of
threads that can be used for preprocessing single-cell data.

The option is designed to parallelize any transformation applied at
individual samples.

Fix an incorrect usage of samples when ignoreSamplesLackingData is set
to true and some samples are missing data.

Fix missing close() on the loader when raising a
NoSingleCellDataFoundException if no data is found on-disk.

Default to 4 threads.

Update completion scripts.