Update commons-lang3 to 3.18.0
Replace all usage of RandomStringUtils to the insecure one in tests.

Use a secure (and strong) random generator for generating a password.
Use secure random generators in other places.

This has to be done because Velocity uses APIs from 3.18 that do not
exist in 3.14 and that triggered an error when attempting to process
more sophisticated templates.