web: Allow deleting QTs, CTAs and CLCs from the UI
Add an isAggregated field to the QuantitationType model to keep track of
QTs that contain aggregated data.

Remove the deleteSingleCellAggregate CLI now that we can detect
aggregated QTs and delete them with deleteRawData.

Indicate which cell types/CTA are mismatched

Codify errors messages in ExpressionExperimentEditController

web: Indicate if the cell type factor was kept, created or replaced